Fix bug: HandleBlockMerge() when f != a.
Let x and z be the children of a. We were calling Detach(x, a) and then
replacing the leaf of the merged blocks with z. This can cause a problem
if there is a group above a that attaches to some other block under z. OOps.

Instead, we simply move x into a's position and kill a. This leaves z
dangling. Now replace the leaf with z. Effectively the same thing, but
avoids the sanity checks and further detaches that Detach() causes.